Split out tooltip component; linting
