+++ js/react/components/GraphContainer.js
@@ -6,7 +6,7 @@ import { Nav, NavItem, Grid, Col } from 'react-bootstrap'
 import moment from 'moment'
 import { LineChart, Line, Tooltip, ResponsiveContainer, XAxis, YAxis } from 'recharts'
 import VerticalLabel from './VerticalLabel'
-import { gameFromTime } from '../utils'
+import GamesTooltip from './GamesTooltip'
 
 const GRAPHS = [
   { name: 'Viewers', key: 'v' },
@@ -21,34 +21,12 @@ const GRAPHS = [
   { name: 'Twitch Emotes per minute', key: 'e' }
 ]
 
-const CustomToolTip = (props) => {
-  const { payload, label, active, schedule } = props
-  
-  if (!active || !payload){
-    return null
-  }
-
-  const payloadObj = payload[0].payload
-  const payloadProps = payload[0]
-  const dataKey = payloadProps.dataKey
-
-  const game = gameFromTime(schedule, label)
-
-  return (
-    <div className='gdq-tooltip'>
-      <div className='tool-game'>{game.name}</div>
-      <div className='tool-date'>{moment(label).format("ddd, MMM Do YYYY, h:mm a")}</div>
-      <div className='tool-primary'>{payloadProps.name}: {payloadObj[dataKey]}</div>
-      <div className='tool-footer'>Baz</div>
-    </div>
-  )
-}
-
 class GraphContainer extends React.Component {
   static propTypes = {
     setCurrentSeries: PropTypes.func.isRequired,
     activeSeries: PropTypes.number.isRequired,
-    timeseries: PropTypes.array.isRequired
+    timeseries: PropTypes.array.isRequired,
+    schedule: PropTypes.array.isRequired
   }
 
   constructor (props) {
@@ -126,8 +104,8 @@ class GraphContainer extends React.Component {
                   dot={false}
                   activeDot />
                 <Tooltip
-                  content={<CustomToolTip schedule={this.props.schedule} />}
-                  animationDuration={0}/>
+                  content={<GamesTooltip schedule={this.props.schedule} />}
+                  animationDuration={0} />
                 <XAxis
                   dataKey='time'
                   axisLine={{stroke: '#ddd'}}
